working with manifold learning elements dynamically
